app/requirements.txt
click==8.2.1
fastapi==0.115.12
h11==0.16.0
httpx==0.28.1
langchain==0.3.25
numpy==1.26.4
openai==1.84.0
packaging==24.2
pydantic==2.11.5
python-dotenv==1.1.0
requests==2.32.3
starlette==0.46.2
typing_extensions==4.14.0
uvicorn==0.34.3
psycopg2-binary==2.9.10
langchain-community==0.3.25
langchain-openai==0.3.23
pgvector==0.4.1
PyJWT==2.10.1
sentence-transformers==2.6.1
transformers==4.39.3
scikit-learn==1.4.2
scipy==1.13.1

pyahocorasick
